CallEventBase -> CallEvent; pack up all call events
These are small enough to comfortably reside in a single translation
unit.
