+++ lib/eventitem.h
@@ -5,6 +5,7 @@
 
 #include "quotient_common.h"
 
+#include "events/callevents.h"
 #include "events/filesourceinfo.h"
 #include "events/stateevent.h"
 
@@ -101,9 +102,9 @@ inline const StateEventBase* EventItemBase::viewAs<StateEventBase>() const
 }
 
 template <>
-inline const CallEventBase* EventItemBase::viewAs<CallEventBase>() const
+inline const CallEvent* EventItemBase::viewAs<CallEvent>() const
 {
-    return evt->isCallEvent() ? weakPtrCast<const CallEventBase>(evt) : nullptr;
+    return evt->is<CallEvent>() ? weakPtrCast<const CallEvent>(evt) : nullptr;
 }
 
 class QUOTIENT_API PendingEventItem : public EventItemBase {
